+++ releasenotes.md
@@ -1,4 +1,5 @@
 Whats New in v10.0
 --------------------------
+- Added Inventory for finished Compounds
 - Minor improvements in formula attachments section
 - For full details please refer to the CHANGELOG